refactor: separate protocol handler url tools
